+++ Backend/.gitignore
@@ -1,2 +1,2 @@
-node_modules/
 .env
+node_modules/
