+++ Backend/logs/app-2024-09-15.log
@@ -8,3 +8,22 @@
 2024-09-15 12:43:28 [INFO]: Request from Guest at ::1 - GET /api/v1/agent/get-forms - 2024-09-15T09:43:28.496Z
 2024-09-15 12:43:43 [INFO]: Request from Guest at ::1 - GET /api/v1/agent/get-forms?tags=su - 2024-09-15T09:43:43.056Z
 2024-09-15 12:43:50 [INFO]: Request from Guest at ::1 - GET /api/v1/agent/get-forms?tags=not%20so%20nice - 2024-09-15T09:43:50.979Z
+2024-09-15 22:01:53 [INFO]: Request from Guest at ::1 - POST /api/v1/auth/signup/organization - 2024-09-15T19:01:53.503Z
+2024-09-15 22:01:53 [ERROR]: Error in request from Guest at ::1 - POST /api/v1/auth/signup/organization - 2024-09-15T19:01:53.982Z - Error: Validation error
+2024-09-15 22:02:05 [INFO]: Request from Guest at ::1 - POST /api/v1/auth/signup/organization - 2024-09-15T19:02:05.201Z
+2024-09-15 22:02:37 [INFO]: Request from Guest at ::1 - POST /api/v1/orgAdmin/create_form - 2024-09-15T19:02:37.778Z
+2024-09-15 22:03:32 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:03:32.199Z
+2024-09-15 22:03:32 [ERROR]: Error in request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:03:32.252Z - Error: User doesn't exist anymore
+2024-09-15 22:03:39 [INFO]: Request from Guest at ::1 - POST /api/v1/auth/signup/agent - 2024-09-15T19:03:39.031Z
+2024-09-15 22:03:57 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:03:57.226Z
+2024-09-15 22:04:36 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:04:36.137Z
+2024-09-15 22:05:17 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:05:17.514Z
+2024-09-15 22:06:45 [INFO]: Request from Guest at ::1 - POST /api/v1/auth/signup/agent - 2024-09-15T19:06:45.425Z
+2024-09-15 22:07:00 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:07:00.579Z
+2024-09-15 22:07:36 [INFO]: Request from Guest at ::1 - POST /api/v1/auth/signup/agent - 2024-09-15T19:07:36.616Z
+2024-09-15 22:07:48 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:07:48.798Z
+2024-09-15 22:17:44 [INFO]: Request from Guest at ::1 - POST /api/v1/auth/signup/agent - 2024-09-15T19:17:44.956Z
+2024-09-15 22:18:01 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:18:01.646Z
+2024-09-15 22:18:05 [ERROR]: Error in request from undefined at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:18:05.820Z - Error: Invalid image file
+2024-09-15 22:57:43 [INFO]: Request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:57:43.319Z
+2024-09-15 22:58:23 [INFO]: Request from Guest at ::1 - GET /api/v1/orgAdmin/get-responses/3/ - 2024-09-15T19:58:23.638Z
