+++ Backend/logs/error-2024-09-15.log
@@ -1,0 +1,3 @@
+2024-09-15 22:01:53 [ERROR]: Error in request from Guest at ::1 - POST /api/v1/auth/signup/organization - 2024-09-15T19:01:53.982Z - Error: Validation error
+2024-09-15 22:03:32 [ERROR]: Error in request from Guest at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:03:32.252Z - Error: User doesn't exist anymore
+2024-09-15 22:18:05 [ERROR]: Error in request from undefined at ::1 - POST /api/v1/agent/submitForm - 2024-09-15T19:18:05.820Z - Error: Invalid image file
